+++ tests/integration/test_payment_method.py
@@ -614,6 +614,143 @@ class TestPaymentMethod(unittest.TestCase):
         found_paypal_account = PayPalAccount.find(token)
         self.assertFalse(found_paypal_account is None)
 
+    def test_create_payment_method_with_account_type_debit(self):
+        config = Configuration.instantiate()
+        customer_id = Customer.create().customer.id
+        client_token = json.loads(TestHelper.generate_decoded_client_token())
+        authorization_fingerprint = client_token["authorizationFingerprint"]
+        http = ClientApiHttp(config, {
+            "authorization_fingerprint": authorization_fingerprint,
+            "shared_customer_identifier": "fake_identifier",
+            "shared_customer_identifier_type": "testing",
+        })
+        status_code, nonce = http.get_credit_card_nonce({
+            "number": CreditCardNumbers.Hiper,
+            "expirationMonth": "11",
+            "expirationYear": "2099",
+        })
+        self.assertTrue(status_code == 201)
+
+        result = PaymentMethod.create({
+            "customer_id": customer_id,
+            "payment_method_nonce": nonce,
+            "options": {
+                "verify_card": "true",
+                "verification_merchant_account_id": TestHelper.hiper_brl_merchant_account_id,
+                "verification_amount": "1.02",
+                "verification_account_type": "debit",
+            },
+        })
+
+        self.assertTrue(result.is_success)
+        self.assertEqual("debit", result.payment_method.verifications[0]["credit_card"]["account_type"])
+
+    def test_create_payment_method_with_account_type_credit(self):
+        config = Configuration.instantiate()
+        customer_id = Customer.create().customer.id
+        client_token = json.loads(TestHelper.generate_decoded_client_token())
+        authorization_fingerprint = client_token["authorizationFingerprint"]
+        http = ClientApiHttp(config, {
+            "authorization_fingerprint": authorization_fingerprint,
+            "shared_customer_identifier": "fake_identifier",
+            "shared_customer_identifier_type": "testing",
+        })
+        status_code, nonce = http.get_credit_card_nonce({
+            "number": CreditCardNumbers.Hiper,
+            "expirationMonth": "11",
+            "expirationYear": "2099",
+        })
+        self.assertTrue(status_code == 201)
+
+        result = PaymentMethod.create({
+            "customer_id": customer_id,
+            "payment_method_nonce": nonce,
+            "options": {
+                "verify_card": "true",
+                "verification_merchant_account_id": TestHelper.hiper_brl_merchant_account_id,
+                "verification_amount": "1.02",
+                "verification_account_type": "credit",
+            },
+        })
+
+        self.assertTrue(result.is_success)
+        self.assertEqual("credit", result.payment_method.verifications[0]["credit_card"]["account_type"])
+
+    def test_create_credit_card_with_account_type_debit(self):
+        customer = Customer.create().customer
+        result = CreditCard.create({
+            "customer_id": customer.id,
+            "number": CreditCardNumbers.Hiper,
+            "expiration_date": "05/2009",
+            "cvv": "100",
+            "cardholder_name": "John Doe",
+            "options": {
+                "verification_merchant_account_id": TestHelper.hiper_brl_merchant_account_id,
+                "verification_account_type": "debit",
+                "verify_card": True,
+            }
+        })
+        self.assertTrue(result.is_success)
+        self.assertEqual("debit", result.credit_card.verifications[0]["credit_card"]["account_type"])
+
+    def test_create_credit_card_with_account_type_credit(self):
+        customer = Customer.create().customer
+        result = CreditCard.create({
+            "customer_id": customer.id,
+            "number": CreditCardNumbers.Hiper,
+            "expiration_date": "05/2009",
+            "cvv": "100",
+            "cardholder_name": "John Doe",
+            "options": {
+                "verification_merchant_account_id": TestHelper.hiper_brl_merchant_account_id,
+                "verification_account_type": "credit",
+                "verify_card": True,
+            }
+        })
+        self.assertTrue(result.is_success)
+        self.assertEqual("credit", result.credit_card.verifications[0]["credit_card"]["account_type"])
+
+    def test_create_with_usupported_merchant_account(self):
+        customer = Customer.create().customer
+        result = CreditCard.create({
+            "customer_id": customer.id,
+            "number": CreditCardNumbers.Visa,
+            "expiration_date": "05/2009",
+            "cvv": "100",
+            "cardholder_name": "John Doe",
+            "options": {
+                "verification_account_type": "debit",
+                "verify_card": True,
+            }
+        })
+        self.assertFalse(result.is_success)
+
+        errors = result.errors.for_object("credit_card").for_object("options").on("verification_account_type")
+
+        self.assertEqual(1, len(errors))
+        self.assertEqual(ErrorCodes.CreditCard.VerificationAccountTypeNotSupported, errors[0].code)
+
+    def test_create_with_invalid_account_type(self):
+        customer = Customer.create().customer
+        result = CreditCard.create({
+            "customer_id": customer.id,
+            "number": CreditCardNumbers.Hiper,
+            "expiration_date": "05/2009",
+            "cvv": "100",
+            "cardholder_name": "John Doe",
+            "options": {
+                "verification_merchant_account_id": TestHelper.hiper_brl_merchant_account_id,
+                "verification_account_type": "invalid",
+                "verify_card": True,
+            }
+        })
+        self.assertFalse(result.is_success)
+
+        errors = result.errors.for_object("credit_card").for_object("options").on("verification_account_type")
+
+        self.assertEqual(1, len(errors))
+        self.assertEqual(ErrorCodes.CreditCard.VerificationAccountTypeIsInvald, errors[0].code)
+
     def test_find_returns_an_abstract_payment_method(self):
         customer_id = Customer.create().customer.id
         result = PaymentMethod.create({
@@ -718,6 +855,101 @@ class TestPaymentMethod(unittest.TestCase):
         self.assertTrue(updated_credit_card.bin == CreditCardNumbers.MasterCard[:6])
         self.assertTrue(updated_credit_card.last_4 == CreditCardNumbers.MasterCard[-4:])
         self.assertTrue(updated_credit_card.expiration_date == "06/2013")
+
+    def test_update_credit_cards_with_account_type_credit(self):
+        customer = Customer.create().customer
+        result = CreditCard.create({
+            "customer_id": customer.id,
+            "number": CreditCardNumbers.Visa,
+            "expiration_date": "05/2009",
+            "cvv": "100",
+            "cardholder_name": "John Doe",
+        })
+        update_result = PaymentMethod.update(result.credit_card.token, {
+            "cardholder_name": "New Holder",
+            "cvv": "456",
+            "number": CreditCardNumbers.Hiper,
+            "expiration_date": "06/2013",
+            "options": {
+                "verification_merchant_account_id": TestHelper.hiper_brl_merchant_account_id,
+                "verification_account_type": "credit",
+                "verify_card": True,
+            }
+        })
+        self.assertTrue(update_result.is_success)
+        self.assertEqual("credit", update_result.payment_method.verification.credit_card["account_type"])
+
+    def test_update_credit_cards_with_account_type_debit(self):
+        customer = Customer.create().customer
+        result = CreditCard.create({
+            "customer_id": customer.id,
+            "number": CreditCardNumbers.Visa,
+            "expiration_date": "05/2009",
+            "cvv": "100",
+            "cardholder_name": "John Doe",
+        })
+        update_result = PaymentMethod.update(result.credit_card.token, {
+            "cardholder_name": "New Holder",
+            "cvv": "456",
+            "number": CreditCardNumbers.Hiper,
+            "expiration_date": "06/2013",
+            "options": {
+                "verification_merchant_account_id": TestHelper.hiper_brl_merchant_account_id,
+                "verification_account_type": "debit",
+                "verify_card": True,
+            }
+        })
+        self.assertTrue(update_result.is_success)
+        self.assertEqual("debit", update_result.payment_method.verification.credit_card["account_type"])
+
+    def test_update_credit_cards_with_invalid_account_type(self):
+        customer = Customer.create().customer
+        result = CreditCard.create({
+            "customer_id": customer.id,
+            "number": CreditCardNumbers.Visa,
+            "expiration_date": "05/2009",
+            "cvv": "100",
+            "cardholder_name": "John Doe",
+        })
+        update_result = PaymentMethod.update(result.credit_card.token, {
+            "cardholder_name": "New Holder",
+            "cvv": "456",
+            "number": CreditCardNumbers.Hiper,
+            "expiration_date": "06/2013",
+            "options": {
+                "verification_merchant_account_id": TestHelper.hiper_brl_merchant_account_id,
+                "verification_account_type": "invalid",
+                "verify_card": True,
+            }
+        })
+        self.assertFalse(update_result.is_success)
+        errors = update_result.errors.for_object("credit_card").for_object("options").on("verification_account_type")
+        self.assertEqual(1, len(errors))
+        self.assertEqual(ErrorCodes.CreditCard.VerificationAccountTypeIsInvald, errors[0].code)
+
+    def test_update_credit_cards_with_unsupported_merchant_account(self):
+        customer = Customer.create().customer
+        result = CreditCard.create({
+            "customer_id": customer.id,
+            "number": CreditCardNumbers.Visa,
+            "expiration_date": "05/2009",
+            "cvv": "100",
+            "cardholder_name": "John Doe",
+        })
+        update_result = PaymentMethod.update(result.credit_card.token, {
+            "cardholder_name": "New Holder",
+            "cvv": "456",
+            "number": CreditCardNumbers.Visa,
+            "expiration_date": "06/2013",
+            "options": {
+                "verification_account_type": "debit",
+                "verify_card": True,
+            }
+        })
+        self.assertFalse(update_result.is_success)
+        errors = update_result.errors.for_object("credit_card").for_object("options").on("verification_account_type")
+        self.assertEqual(1, len(errors))
+        self.assertEqual(ErrorCodes.CreditCard.VerificationAccountTypeNotSupported, errors[0].code)
 
     def test_update_creates_a_new_billing_address_by_default(self):
         customer_id = Customer.create().customer.id