+++ tests/unit/test_payment_method_gateway.py
@@ -34,6 +34,7 @@ class TestPaymentMethodGateway(unittest.TestCase):
                     "verification_merchant_account_id",
                     "verify_card",
                     "verification_amount",
+                    "verification_account_type",
                     {
                         "adyen":[
                             "overwrite_brand",
@@ -96,6 +97,7 @@ class TestPaymentMethodGateway(unittest.TestCase):
                     "verify_card",
                     "verification_amount",
                     "verification_merchant_account_id",
+                    "verification_account_type",
                     "venmo_sdk_session",
                     {
                         "adyen":[